Use the implementation instead of the accidental infinite loop
